Switched to using floats instead of long doubles

+++ main.hpp
@@ -1,6 +1,6 @@
 #ifndef COURSEWORKONE_MAIN_HPP
 #define COURSEWORKONE_MAIN_HPP
-using float_type = long double;
+using float_type = float;
 void presentPart1 ();
 void presentPart3 ();
 #endif //COURSEWORKONE_MAIN_HPP
